requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
rank-bm25>=0.2.2
sentence-transformers>=3.0
faiss-cpu>=1.8.0
pyyaml>=6.0
matplotlib>=3.9
datasets>=3.0
